+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-## 0.19.2
-- solargraph._openDocumentUrl escapes hashes
+## 0.19.2 - January 28, 2019
+- solargraph._openDocumentUrl escapes special characters
+- Show error message when solargraph gem is not installed
 
 ## 0.19.1 - December 28, 2018
 - Fixed FileSystemWatcher glob pattern